+++ src/main/java/com/nyc/hosp/model/HospuserDTO.java
@@ -7,7 +7,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 public class HospuserDTO {
 
-    private Integer userId;
+    private Long userId;
 
     @Size(max = 100)
     //@Pattern()
@@ -40,11 +40,11 @@ public class HospuserDTO {
 
     private Integer role;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(final Integer userId) {
+    public void setUserId(final Long userId) {
         this.userId = userId;
     }
